Show predicted class
